act backend

+++ src/main/java/pe/edu/upeu/Spring/Controller/ProgramaController.java
@@ -36,15 +36,13 @@ public class ProgramaController {
     @Autowired
     private ProgramaService programaService;
 
-
     @ApiOperation(value = "Lista de taller")
     @GetMapping()
     public ResponseEntity<?> findAll() {
         HashMap<String, Object> result = new HashMap<>();
-result.put("success", true);
-result.put("message", "Lista de programas");
-result.put("data", programaService.findAll());
-
+        result.put("success", true);
+        result.put("message", "Lista de programas");
+        result.put("data", programaService.findAll());
 
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
@@ -58,19 +56,54 @@ result.put("data", programaService.findAll());
 
     @ApiOperation(value = "Elimina un taller")
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        programaService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe programa con id:" + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+
+        } else {
+            programaService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Registrado eliminado correctamente");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
+
     }
 
-    @ApiOperation(value = "Crea un taller")
-    @PostMapping("/save")
-    public Programa save(@RequestBody Programa programa) {
-        return programaService.save(programa);
+    @ApiOperation(value = "Crea una programa")
+    @PostMapping
+    public ResponseEntity<?> save(@RequestBody Programa programa) {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Registrado creado exitosamente");
+        result.put("data", programaService.save(programa));
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
-    @ApiOperation(value = "Modifica una persona")
-    @PutMapping("/update")
-    public Programa update(@RequestBody Programa programa) {
-        return programaService.save(programa);
-    }
-}
+
+
+
+    @ApiOperation(value = "Modifica una programa")
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Programa programa) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe registro con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            programa.setProgId(id);
+            programaService.save(programa);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", programa);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }}
